+++ packages/admin/dashboard/src/routes/orders/order-create-return/components/return-create-form/return-item.tsx
@@ -53,10 +53,10 @@ function ReturnItem({
               <Text className="txt-small" as="span" weight="plus">
                 {item.title}{" "}
               </Text>
-              {item.variant.sku && <span>({item.variant.sku})</span>}
+              {item.variant_sku && <span>({item.variant_sku})</span>}
             </div>
             <Text as="div" className="text-ui-fg-subtle txt-small">
-              {item.variant.product.title}
+              {item.product_title}
             </Text>
           </div>
         </div>