+++ packages/core/core-flows/src/cart/workflows/update-cart.ts
@@ -34,18 +34,32 @@ export const updateCartWorkflow = createWorkflow(
     const cartToUpdate = useRemoteQueryStep({
       entry_point: "cart",
       variables: { id: input.id },
-      fields: ["id", "shipping_address.*", "region.*", "region.countries.*"],
+      fields: [
+        "id",
+        "email",
+        "customer_id",
+        "shipping_address.*",
+        "region.*",
+        "region.countries.*",
+      ],
       list: false,
       throw_if_key_not_found: true,
     }).config({ name: "get-cart" })
 
-    const [salesChannel, customerData] = parallelize(
+    const customerDataInput = transform({ input, cartToUpdate }, (data) => {
+      return {
+        customer_id: data.cartToUpdate.customer_id,
+        email: data.input.email ?? data.cartToUpdate.email,
+      }
+    })
+
+    const [salesChannel, customer] = parallelize(
       findSalesChannelStep({
         salesChannelId: input.sales_channel_id,
       }),
       findOrCreateCustomerStep({
-        customerId: input.customer_id,
-        email: input.email,
+        customerId: customerDataInput.customer_id,
+        email: customerDataInput.email,
       })
     )
 
@@ -66,7 +80,13 @@ export const updateCartWorkflow = createWorkflow(
     })
 
     const cartInput = transform(
-      { input, region, customerData, salesChannel, cartToUpdate },
+      {
+        input,
+        region,
+        customer,
+        salesChannel,
+        cartToUpdate,
+      },
       (data) => {
         const {
           promo_codes,
@@ -116,13 +136,16 @@ export const updateCartWorkflow = createWorkflow(
           }
         }
 
-        if (
-          isDefined(updateCartData.customer_id) ||
-          isDefined(updateCartData.email)
-        ) {
-          data_.customer_id = data.customerData.customer?.id || null
-          data_.email =
-            data.input?.email ?? (data.customerData.customer?.email || null)
+        if (isDefined(updateCartData.email) && data.customer?.customer) {
+          const currentCustomer = data.customer.customer!
+          data_.customer_id = currentCustomer.id
+
+          // registered customers can update the cart email
+          if (currentCustomer.has_account) {
+            data_.email = updateCartData.email
+          } else {
+            data_.email = data.customer.email
+          }
         }
 
         if (isDefined(updateCartData.sales_channel_id)) {
@@ -133,6 +156,7 @@ export const updateCartWorkflow = createWorkflow(
       }
     )
 
+    /*
     when({ cartInput }, ({ cartInput }) => {
       return isDefined(cartInput.customer_id) || isDefined(cartInput.email)
     }).then(() => {
@@ -141,6 +165,7 @@ export const updateCartWorkflow = createWorkflow(
         data: { id: input.id },
       }).config({ name: "emit-customer-updated" })
     })
+    */
 
     when({ input, cartToUpdate }, ({ input, cartToUpdate }) => {
       return (