+++ packages/modules/providers/payment-stripe/src/types/index.ts
@@ -43,4 +43,5 @@ export const PaymentProviderKeys = {
   GIROPAY: "stripe-giropay",
   IDEAL: "stripe-ideal",
   PRZELEWY_24: "stripe-przelewy24",
+  PROMPT_PAY : "stripe-promptpay",
 }