+++ packages/admin/admin-bundler/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @medusajs/admin-bundler
 
+## 2.0.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @medusajs/dashboard@2.0.3
+  - @medusajs/admin-shared@2.0.3
+  - @medusajs/admin-vite-plugin@2.0.3
+
 ## 2.0.2
 
 ### Patch Changes