+++ packages/admin/admin-shared/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @medusajs/admin-shared
 
+## 2.0.3
+
 ## 2.0.2
 
 ## 2.0.1