+++ packages/admin/dashboard/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @medusajs/dashboard
 
+## 2.0.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @medusajs/admin-shared@2.0.3
+  - @medusajs/js-sdk@2.0.3
+  - @medusajs/icons@2.0.3
+
 ## 2.0.2
 
 ### Patch Changes