+++ packages/cli/create-medusa-app/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 2.0.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @medusajs/telemetry@2.0.3
+
 ## 2.0.2
 
 ### Patch Changes