+++ packages/modules/workflow-engine-redis/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @medusajs/workflow-engine-redis
 
+## 2.0.3
+
 ## 2.0.2
 
 ## 2.0.1